removed most unimportant parts of the dataset
